Move player main command into folder and refactor
